+++ DirectShowFilters/MPAudioRenderer/source/TimeStretchFilter.h
@@ -48,28 +48,28 @@ public:
 
   /// Sets new rate control value. Normal rate = 1.0, smaller values
   /// represent slower rate, larger faster rates.
-  void setRate(double newRate);
+  void setRate(float newRate);
 
   /// Sets new tempo control value. Normal tempo = 1.0, smaller values
   /// represent slower tempo, larger faster tempo.
-  void setTempo(double newTempo, double newAdjustment);
+  void setTempo(float newTempo, float newAdjustment);
 
   /// Sets new rate control value as a difference in percents compared
   /// to the original rate (-50 .. +100 %)
-  void setRateChange(double newRate);
+  void setRateChange(float newRate);
 
   /// Sets new tempo control value as a difference in percents compared
   /// to the original tempo (-50 .. +100 %)
-  void setTempoChange(double newTempo);
+  void setTempoChange(float newTempo);
 
   /// Sets pitch change in octaves compared to the original pitch  
   /// (-1.00 .. +1.00)
-  void setPitchOctaves(double newPitch);
+  void setPitchOctaves(float newPitch);
 
   /// Sets pitch change in semi-tones compared to the original pitch
   /// (-12 .. +12)
   void setPitchSemiTones(int newPitch);
-  void setPitchSemiTones(double newPitch);
+  void setPitchSemiTones(float newPitch);
 
   /// Sets sample rate.
   void setSampleRate(uint srate);
@@ -131,7 +131,7 @@ protected:
   bool putSamplesInternal(const short *inBuffer, long inSamples);
   uint receiveSamplesInternal(short *outBuffer, uint maxSamples);
 
-  void setTempoInternal(double newTempo, double newAdjustment);
+  void setTempoInternal(float newTempo, float newAdjustment);
 
 // Internal implementation
 private:
